+++ src/main/java/com/luxoft/java7se/exercise10/exception/OverDraftLimitExceededException.java
@@ -3,7 +3,7 @@ package com.luxoft.java7se.exercise10.exception;
 import java.math.BigDecimal;
 
 public class OverDraftLimitExceededException extends NotEnoughFundsException {
-    public OverDraftLimitExceededException(BigDecimal amount) {
+    public OverDraftLimitExceededException(double amount) {
         super(amount);
     }
 }
